similar_images/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "similar-images"
version = "0.1.0"
description = "Find similar images by perceptual hash with configurable level (strict/medium/loose)."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
keywords = ["image", "similarity", "perceptual-hash", "deduplication"]
dependencies = [
    "Pillow>=9.0",
    "imagehash>=4.3",
    "tqdm>=4.65",
]

[project.scripts]
find-similar = "similar_images.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["similar_images*"]
